examples/02-micro_vm-python-flask/requirements.txt
flask[async]
nats-py
msgpack